backend/requirements.txt
django>=5.0,<6.1
djangorestframework>=3.16,<3.17
channels>=4.1,<4.2
channels-redis>=4.2,<4.3
django-cors-headers>=4.9,<5.0
python-dotenv>=1.0,<2.0
psycopg[binary]>=3.1,<3.3
drf-spectacular>=0.27,<0.28
gunicorn>=22.0,<23.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
requests>=2.31,<3.0
